+++ src/Notifications/Notifications/UnhealthyBackupWasFoundNotification.php
@@ -5,14 +5,17 @@ namespace Spatie\Backup\Notifications\Notifications;
 use Illuminate\Notifications\Messages\MailMessage;
 use Illuminate\Notifications\Messages\SlackAttachment;
 use Illuminate\Notifications\Messages\SlackMessage;
-use Spatie\Backup\Events\UnhealthyBackupWasFound as UnhealthyBackupWasFoundEvent;
+use Spatie\Backup\Events\UnhealthyBackupWasFound;
 use Spatie\Backup\Notifications\BaseNotification;
+use Spatie\Backup\Notifications\Channels\Discord\DiscordMessage;
 use Spatie\Backup\Tasks\Monitor\HealthCheckFailure;
 
-class UnhealthyBackupWasFound extends BaseNotification
+class UnhealthyBackupWasFoundNotification extends BaseNotification
 {
-    /** @var \Spatie\Backup\Events\UnhealthyBackupWasFound */
-    protected $event;
+    public function __construct(
+        public UnhealthyBackupWasFound $event,
+    ) {
+    }
 
     public function toMail(): MailMessage
     {
@@ -70,6 +73,29 @@ class UnhealthyBackupWasFound extends BaseNotification
         return $slackMessage;
     }
 
+    public function toDiscord(): DiscordMessage
+    {
+        $discordMessage = (new DiscordMessage())
+            ->error()
+            ->from(config('backup.notifications.discord.username'), config('backup.notifications.discord.avatar_url'))
+            ->title(
+                trans('backup::notifications.unhealthy_backup_found_subject_title', [
+                    'application_name' => $this->applicationName(),
+                    'problem' => $this->problemDescription(),
+                ])
+            )->fields($this->backupDestinationProperties()->toArray());
+
+        if ($this->failure()->wasUnexpected()) {
+            $discordMessage
+                ->fields(['Health Check' => $this->failure()->healthCheck()->name()])
+                ->fields([
+                    trans('backup::notifications.exception_message_title') => $this->failure()->exception()->getMessage(),
+                ]);
+        }
+
+        return $discordMessage;
+    }
+
     protected function problemDescription(): string
     {
         if ($this->failure()->wasUnexpected()) {
@@ -82,12 +108,5 @@ class UnhealthyBackupWasFound extends BaseNotification
     protected function failure(): HealthCheckFailure
     {
         return $this->event->backupDestinationStatus->getHealthCheckFailure();
-    }
-
-    public function setEvent(UnhealthyBackupWasFoundEvent $event)
-    {
-        $this->event = $event;
-
-        return $this;
     }
 }