+++ src/Tasks/Backup/DbDumperFactory.php
@@ -15,14 +15,19 @@ use Spatie\DbDumper\DbDumper;
 
 class DbDumperFactory
 {
-    protected static $custom = [];
+    protected static array $custom = [];
 
     public static function createFromConnection(string $dbConnectionName): DbDumper
     {
         $parser = new ConfigurationUrlParser();
+
+        if (config("database.connections.{$dbConnectionName}") === null) {
+            throw CannotCreateDbDumper::unsupportedDriver($dbConnectionName);
+        }
+
         try {
             $dbConfig = $parser->parseConfiguration(config("database.connections.{$dbConnectionName}"));
-        } catch (Exception $e) {
+        } catch (Exception) {
             throw CannotCreateDbDumper::unsupportedDriver($dbConnectionName);
         }
 
@@ -53,6 +58,10 @@ class DbDumperFactory
 
         if (isset($dbConfig['dump'])) {
             $dbDumper = static::processExtraDumpParameters($dbConfig['dump'], $dbDumper);
+        }
+
+        if (isset($dbConfig['unix_socket'])) {
+            $dbDumper = $dbDumper->setSocket($dbConfig['unix_socket']);
         }
 
         return $dbDumper;
@@ -122,8 +131,6 @@ class DbDumperFactory
     protected static function determineValidMethodName(DbDumper $dbDumper, string $methodName): string
     {
         return collect([$methodName, 'set'.ucfirst($methodName)])
-            ->first(function (string $methodName) use ($dbDumper) {
-                return method_exists($dbDumper, $methodName);
-            }, '');
+            ->first(fn (string $methodName) => method_exists($dbDumper, $methodName), '');
     }
 }