+++ src/Tasks/Backup/Zip.php
@@ -8,23 +8,23 @@ use ZipArchive;
 
 class Zip
 {
-    /** @var \ZipArchive */
-    protected $zipFile;
+    protected ZipArchive $zipFile;
 
-    /** @var int */
-    protected $fileCount = 0;
+    protected int $fileCount = 0;
 
-    /** @var string */
-    protected $pathToZip;
+    protected string $pathToZip;
 
     public static function createForManifest(Manifest $manifest, string $pathToZip): self
     {
+        $relativePath = config('backup.backup.source.files.relative_path') ?
+            rtrim(config('backup.backup.source.files.relative_path'), DIRECTORY_SEPARATOR) . DIRECTORY_SEPARATOR : false;
+
         $zip = new static($pathToZip);
 
         $zip->open();
 
         foreach ($manifest->files() as $file) {
-            $zip->add($file, self::determineNameOfFileInZip($file, $pathToZip));
+            $zip->add($file, self::determineNameOfFileInZip($file, $relativePath));
         }
 
         $zip->close();
@@ -32,14 +32,12 @@ class Zip
         return $zip;
     }
 
-    protected static function determineNameOfFileInZip(string $pathToFile, string $pathToZip)
+    protected static function determineNameOfFileInZip(string $pathToFile, string $relativePath)
     {
-        $zipDirectory = pathinfo($pathToZip, PATHINFO_DIRNAME);
+        $fileDirectory = pathinfo($pathToFile, PATHINFO_DIRNAME) . DIRECTORY_SEPARATOR;
 
-        $fileDirectory = pathinfo($pathToFile, PATHINFO_DIRNAME);
-
-        if (Str::startsWith($fileDirectory, $zipDirectory)) {
-            return str_replace($zipDirectory, '', $pathToFile);
+        if ($relativePath && $relativePath != DIRECTORY_SEPARATOR && Str::startsWith($fileDirectory, $relativePath)) {
+            return str_replace($relativePath, '', $pathToFile);
         }
 
         return $pathToFile;
@@ -73,23 +71,17 @@ class Zip
         return Format::humanReadableSize($this->size());
     }
 
-    public function open()
+    public function open(): void
     {
         $this->zipFile->open($this->pathToZip, ZipArchive::CREATE);
     }
 
-    public function close()
+    public function close(): void
     {
         $this->zipFile->close();
     }
 
-    /**
-     * @param string|array $files
-     * @param string $nameInZip
-     *
-     * @return \Spatie\Backup\Tasks\Backup\Zip
-     */
-    public function add($files, string $nameInZip = null): self
+    public function add(string | iterable $files, string $nameInZip = null): self
     {
         if (is_array($files)) {
             $nameInZip = null;
@@ -101,7 +93,7 @@ class Zip
 
         foreach ($files as $file) {
             if (is_dir($file)) {
-                $this->zipFile->addEmptyDir($file);
+                $this->zipFile->addEmptyDir(ltrim($nameInZip ?: $file, DIRECTORY_SEPARATOR));
             }
 
             if (is_file($file)) {