+++ src/Tasks/Monitor/HealthChecks/IsReachable.php
@@ -7,11 +7,11 @@ use Spatie\Backup\Tasks\Monitor\HealthCheck;
 
 class IsReachable extends HealthCheck
 {
-    public function checkHealth(BackupDestination $backupDestination)
+    public function checkHealth(BackupDestination $backupDestination): void
     {
         $this->failUnless(
             $backupDestination->isReachable(),
-            trans('backup::notification.unhealthy_backup_found_not_reachable', [
+            trans('backup::notifications.unhealthy_backup_found_not_reachable', [
                 'error' => $backupDestination->connectionError,
             ])
         );